+++ pages/api/workspaces/[workspaceId]/cocktails/index.tsx
@@ -19,14 +19,19 @@ export default withHttpMethods({
         workspaceId: workspace.id,
       },
       include: {
-        glass: true,
-        garnishes: { include: { garnish: true } },
+        _count: { select: { CocktailRecipeImage: true } },
+        glass: { include: { _count: { select: { GlassImage: true } } } },
+        garnishes: {
+          include: {
+            garnish: { include: { _count: { select: { GarnishImage: true } } } },
+          },
+        },
         steps: {
           include: {
             action: true,
             ingredients: {
               include: {
-                ingredient: true,
+                ingredient: { include: { _count: { select: { IngredientImage: true } } } },
               },
             },
           },
@@ -41,15 +46,15 @@ export default withHttpMethods({
       const result = cocktailRecipes.filter(
         (cocktail) =>
           cocktail.name.toLowerCase().includes(search) ||
-          (cocktail.tags.some((tag) => tag.toLowerCase().includes(search)) && search.length > 3) ||
-          (cocktail.garnishes.some((garnish) => garnish.garnish.name.toLowerCase().includes(search)) && search.length > 3) ||
+          (cocktail.tags.some((tag) => tag.toLowerCase().includes(search)) && search.length >= 3) ||
+          (cocktail.garnishes.some((garnish) => garnish.garnish.name.toLowerCase().includes(search)) && search.length >= 3) ||
           cocktail.steps.some((step) =>
             step.ingredients
               .filter((ingredient) => ingredient.ingredient?.name != undefined)
               .some(
                 (ingredient) =>
-                  (ingredient.ingredient?.name.toLowerCase().includes(search) && search.length > 3) ||
-                  ((ingredient.ingredient?.shortName ?? '').toLowerCase().includes(search) && search.length > 3),
+                  (ingredient.ingredient?.name.toLowerCase().includes(search) && search.length >= 3) ||
+                  ((ingredient.ingredient?.shortName ?? '').toLowerCase().includes(search) && search.length >= 3),
               ),
           ),
       );