+++ docs/CHANGELOG.md
@@ -1,3 +1,26 @@
+# [1.2.0](https://github.com/jo-gross/Cocktail-Manager/compare/v1.1.1...v1.2.0) (2024-02-21)
+
+
+### Bug Fixes
+
+* Changed search condition to shorter length ([8ee37aa](https://github.com/jo-gross/Cocktail-Manager/commit/8ee37aa2013ef78c6bd36b67702486b870915acd))
+* Disabled metro import ([17c5240](https://github.com/jo-gross/Cocktail-Manager/commit/17c5240f0cf378a64082eb0d0dfa47ed12ee2128))
+* Disabled metro import ([7df6a50](https://github.com/jo-gross/Cocktail-Manager/commit/7df6a500863fabd03f117515e4f75b5841fcfe76))
+* Fixed logo theme ([b669e22](https://github.com/jo-gross/Cocktail-Manager/commit/b669e2265f5e4145a132e54c88833b1da7111d35))
+* Fixed theme changer, and added auto-theme option ([2803af0](https://github.com/jo-gross/Cocktail-Manager/commit/2803af0ef4d7f81419e77f43e44298b653bd00d7))
+* Last working version ([0cea5ee](https://github.com/jo-gross/Cocktail-Manager/commit/0cea5ee946910387c0c85d38329675cd64ddeab2))
+* Last working version ([688d0e1](https://github.com/jo-gross/Cocktail-Manager/commit/688d0e1c89d4f22e7c44dc41ee756f188b85d9d8))
+* Made imports more failure tolerant ([a5f372c](https://github.com/jo-gross/Cocktail-Manager/commit/a5f372c48db4c42d5eaead489b940009db2966e7))
+* Made imports more failure tolerant ([58ab6cd](https://github.com/jo-gross/Cocktail-Manager/commit/58ab6cd080046949af16cd6f3cf5521fd1451222))
+* Removed price label if no price exists ([0dec065](https://github.com/jo-gross/Cocktail-Manager/commit/0dec065b83dfabc46f85d9fd8da344724e769fe2))
+
+
+### Features
+
+* Improved image loading, by loading only existing images ([d7536df](https://github.com/jo-gross/Cocktail-Manager/commit/d7536df4cf997c3daea9bcebfa8e510f6d0f78cd))
+* Improved image loading, by loading only existing images ([0278b65](https://github.com/jo-gross/Cocktail-Manager/commit/0278b65e2c5395d9de289651528a55df6d337a97))
+* Made all prices (cocktails, ingredients, garnishes) optional ([d8abd5a](https://github.com/jo-gross/Cocktail-Manager/commit/d8abd5a5d4a600b9f870d52a253a1cd10720de3c))
+
 ## [1.1.1](https://github.com/jo-gross/Cocktail-Manager/compare/v1.1.0...v1.1.1) (2024-02-19)